// src/store/actions/buildings/interfaces.ts
export enum getBuildingsType {
  GET_BUILDINGS = 'GET_BUILDINGS',
  GET_BUILDINGS_FAIL = 'GET_BUILDINGS_FAIL',
  GET_BUILDINGS_SUCCESS = 'GET_BUILDINGS_SUCCESS',
  SHOW_BUILDING = 'SHOW_BUILDING',
  RESET_EDITING = 'RESET_EDITING',
  SHOW_WARNING = 'SHOW_WARNING',
  CANCEL_DELETE = 'CANCEL_DELETE',
  EDIT_BUILDING_FAIL = 'EDIT_BUILDING_FAIL',
  EDIT_BUILDING_SUCCESS = 'EDIT_BUILDING_SUCCESS',
  ADD_BUILDING_FAIL = 'ADD_BUILDING_FAIL',
  RESET_BUILDINGS_ERRORS = 'RESET_BUILDINGS_ERRORS'
}